updated gui of follow line superthin
